pid lst などのpid関係の関数をcmd_pid.cに入れた
